pyproject.toml
[project]
name = "india-insight-agent"
version = "0.1.0"
description = "AI Agent for MoSPI data insights"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "mcp",
    "fastapi",
    "uvicorn",
]

[tool.uv]
dev-dependencies = []
